+++ scrutiny/gui/component_app_interface.py
@@ -9,7 +9,7 @@
 import abc
 
 from scrutiny.gui.core.watchable_registry import WatchableRegistry
-from scrutiny.gui.core.server_manager import ServerManager
+from scrutiny.gui.core.server_manager.server_manager import ServerManager
 
 
 class AbstractComponentAppInterface(abc.ABC):
