+++ scrutiny/gui/dialogs/server_sfd_manager_dialog.py
@@ -18,7 +18,7 @@ from PySide6.QtGui import QCloseEvent, QKeyEvent, QShowEvent, QStandardItemModel
 
 from scrutiny import sdk
 from scrutiny.sdk.client import ScrutinyClient, SFDUploadRequest, SFDDownloadRequest
-from scrutiny.gui.core.server_manager import ServerManager
+from scrutiny.gui.core.server_manager.server_manager  import ServerManager
 from scrutiny.gui.core.persistent_data import gui_persistent_data
 from scrutiny.gui.themes import scrutiny_get_theme
 from scrutiny.gui.widgets.feedback_label import FeedbackLabel
