+++ scrutiny/gui/main_window.py
@@ -43,7 +43,7 @@ from scrutiny.gui.components.locals.hmi.hmi_component import HMIComponent
 from scrutiny.gui.components.globals.metrics.metrics_component import MetricsComponent
 
 from scrutiny.gui.core.user_messages_manager import UserMessagesManager
-from scrutiny.gui.core.server_manager import ServerManager
+from scrutiny.gui.core.server_manager.server_manager import ServerManager
 from scrutiny.gui.core.watchable_registry import WatchableRegistry
 
 from scrutiny import tools
