+++ test/gui/test_dashboard.py
@@ -99,15 +99,6 @@ class MainWindowStub(QWidget):
 class TestDashboard(ScrutinyBaseGuiTest):
     def setUp(self):
         super().setUp()
-        settings = ScrutinyQtGUI.Settings(
-            debug_layout=False,
-            auto_connect=False,
-            opengl_enabled=False,
-            local_server_port=8765,
-            start_local_server=False,
-            theme=SupportedTheme.Default
-        )
-        configure_unit_test_app_settings(settings)
         self.main_window = MainWindowStub()
 
     def test_setup_teardown(self):
