+++ test/gui/test_watch_component.py
@@ -40,16 +40,6 @@ class TestWatchComponent(ScrutinyBaseGuiTest):
     def setUp(self):
         super().setUp()
 
-        settings = ScrutinyQtGUI.Settings(
-            debug_layout=False,
-            auto_connect=False,
-            opengl_enabled=False,
-            local_server_port=8765,
-            start_local_server=False,
-            theme=SupportedTheme.Default
-        )
-        configure_unit_test_app_settings(settings)
-
         self.main_window = MainWindowStub()
         app_interface = DummyAppInterface()
         app_interface.server_manager = self.main_window.get_server_manager()
